Refactor user lists and learning paths hooks out of learning resources (#1842)
* Refactor learning path API hooks out of learning resources

* Refactor user list API hooks out of learning resources. Dead code commented out

* Move post mutation helpers to hooks files

* Remove empty file

* Remove commented code
