+++ frontends/api/src/hooks/programLetters/index.ts
@@ -2,7 +2,7 @@ import { useQuery } from "@tanstack/react-query"
 import programLetters from "./keyFactory"
 
 /**
- * Query is diabled if id is undefined.
+ * Query is disabled if id is undefined.
  */
 const useProgramLettersDetail = (id: string | undefined) => {
   return useQuery({
